requirements.txt
Django>=4.2,<4.3
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-netfields==1.3.1
django-prometheus==2.3.1
django-redis==5.4.0
django-rq==2.8.1
django-tables2==2.6.0
django-taggit==4.0.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
Jinja2>=3.1.2
Markdown==3.5
napalm==4.1.0
packaging==23.2
psycopg[binary,pool]==3.1.12
pyixapi==0.2.1
pynetbox==7.2.0
PyYAML==6.0.1
requests==2.31.0
tzdata==2023.3
napalm-sros>=1.0.1
pycryptodome>=3.15.0